c-s separation
